requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
h11==0.16.0
idna==3.10
numpy==2.3.2
outcome==1.3.0.post0
pandas==2.3.2
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
playwright==1.49.0
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
openai>=1.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
streamlit>=1.28.0
